Fixed a problem where BaseEvents could not be casted to Tests, Assignments or Classes
