+++ src/com/merendabot/university/events/Test.java
@@ -1,5 +1,8 @@
 package com.merendabot.university.events;
 
+import com.merendabot.university.subjects.Subject;
+
+import java.sql.Date;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.time.LocalDate;
@@ -7,11 +10,20 @@ import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 
-public interface Test extends BaseEvent {
+public class Test extends BaseEventClass {
 
-    LocalDate getDueDate();
+    public Test(int id, String name, LocalDate dueDate, LocalTime dueTime, String link, Subject subject) {
+        super(id, EventType.TEST, EventInterval.SINGLE, name, dueDate, (Date) null, dueTime, null, link, subject);
+    }
 
-    LocalTime getDueTime();
+
+    public LocalDate getDueDate() {
+        return getStartDate();
+    }
+
+    public LocalTime getDueTime() {
+        return getStartTime();
+    }
 
     /**
      * Gets all tests from the database.
@@ -19,7 +31,7 @@ public interface Test extends BaseEvent {
      * @return A List of Tests
      * @throws SQLException if an SQL Error occurs
      */
-    static List<Test> getTests() throws SQLException {
+    public static List<Test> getTests() throws SQLException {
         List<Test> tests = new ArrayList<>();
         ResultSet rs = BaseEvent.getEvents(EventType.TEST);
         while (rs.next())
@@ -34,7 +46,7 @@ public interface Test extends BaseEvent {
      * @return A Test object if test exists, null otherwise.
      * @throws SQLException If an SQL Error occurs
      */
-    static Test getTestById(int id) throws SQLException {
+    public static Test getTestById(int id) throws SQLException {
         return (Test) BaseEvent.getEventById(id);
     }
 }
